+++ lending/src/lib/config.ts
@@ -52,7 +52,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "0.7.47",
-  genVersion: "2.492.3",
-  userAgent: "speakeasy-sdk/typescript 0.7.47 2.492.3 1.0.0 ryan-lending",
+  sdkVersion: "0.7.48",
+  genVersion: "2.493.4",
+  userAgent: "speakeasy-sdk/typescript 0.7.48 2.493.4 1.0.0 ryan-lending",
 } as const;